Login/Logout to NavBar and Created Default Home
When the user is not signed in, the home page displays a welcome message and prompts the user to sign in through Google. The dropdown menu will also have a login link, that prompts users to sign in to view the full dropdown menu. (polish front end design later)

When the user is signed in, the dropdown menu will show profile, settings, and sign out and the Home page will default to the spreadsheet page.